changed float to douyble on va_arg from p_flo

+++ 0x10-variadic_functions/3-print_all.c
@@ -36,7 +36,7 @@ void print_int(va_list list)
 
 void print_float(va_list list)
 {
-	printf("%f", va_arg(list, float));
+	printf("%f", va_arg(list, double));
 }
 
 /**
